pyproject.toml
[project]
name = "markit"
version = "0.1.2"
description = "Intelligent document to Markdown conversion tool with LLM enhancement"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "MarkIt Team" }]
keywords = ["markdown", "document", "conversion", "llm", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Utilities",
]

dependencies = [
    # CLI
    "typer>=0.21.0",
    "rich>=14.2.0",
    # Configuration
    "pydantic>=2.12.0",
    "pydantic-settings>=2.12.0",
    # Conversion engines
    "markitdown[all]>=0.1.4",      # Microsoft conversion library (with all format support)
    "pypandoc>=1.15",              # Pandoc Python wrapper
    # PDF processing
    "pymupdf>=1.26.0",             # PDF processing (speed priority)
    "pdfplumber>=0.11.4",          # PDF processing (table friendly)
    "pymupdf4llm>=0.2.0",          # PDF to Markdown for LLM (tables, headings, images)
    "pymupdf-layout>=0.1.0",       # Improved page layout analysis for pymupdf4llm
    # Image processing
    "pillow>=12.0.0",              # Image processing (12.x has built-in SIMD optimization)
    # Async & HTTP
    "anyio>=4.12.0",
    "httpx>=0.28.0",
    # LLM SDK - Official libraries
    "openai>=2.14.0",              # OpenAI official SDK
    "google-genai>=1.56.0",        # Google new unified SDK
    "ollama>=0.6.1",               # Ollama official library
    "anthropic>=0.75.0",           # Anthropic official SDK
    # Text chunking
    "langchain-text-splitters>=1.1.0",  # Mature chunking library
    "tiktoken>=0.12.0",            # OpenAI tokenizer
    # Logging
    "structlog>=25.5.0",
    # Utilities
    "tenacity>=9.1.0",             # Retry
    "python-magic>=0.4.27;sys_platform!='win32'",      # File type detection (Linux/macOS)
    "python-magic-bin>=0.4.14;sys_platform=='win32'",  # Windows compatible
    # Windows-specific: MS Office COM automation
    "pywin32>=310;sys_platform=='win32'",
    "tomlkit>=0.13.3",
    "PyYAML>=6.0.1",
    "questionary>=2.1.1",
    "ruamel-yaml>=0.19.1",
]

[project.optional-dependencies]
# High performance chunking (optional)
performance = [
    "semantic-text-splitter>=0.22.0",  # Rust implementation for high performance
]

# OCR support for scanned PDFs (optional)
ocr = [
    "opencv-python>=4.10.0",           # Required for pymupdf4llm OCR
]

dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0.0",
    "pytest-sugar>=1.1.1",
    "ruff>=0.14.0",
    "pyright>=1.1.391",
    "pre-commit>=4.0.0",
]

[project.scripts]
markit = "markit.cli.main:app"

[project.urls]
Homepage = "https://github.com/Ynewtime/markit"
Documentation = "https://github.com/Ynewtime/markit#readme"
Repository = "https://github.com/Ynewtime/markit"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/markit"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B009",   # do not call getattr with a constant attribute value
    "UP007",  # Use X | None instead of Optional[X] - allow both styles
    "UP046",  # Use type params instead of Generic[T] - allow for compatibility
    "SIM108", # Use ternary operator - allow if-else blocks for readability
    "SIM102", # Nested if - allow for readability in complex conditions
]

[tool.ruff.lint.isort]
known-first-party = ["markit"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "--tb=short -m 'not e2e'"
markers = [
    "e2e: end-to-end tests requiring external services",
    "integration: integration tests",
    "ollama: tests requiring local Ollama service",
    "openai: tests requiring OPENAI_API_KEY",
    "anthropic: tests requiring ANTHROPIC_API_KEY",
    "gemini: tests requiring GOOGLE_API_KEY",
    "openrouter: tests requiring OPENROUTER_API_KEY",
]

[tool.pyright]
pythonVersion = "3.12"
include = ["src/markit"]
typeCheckingMode = "standard"
reportMissingImports = true
reportMissingTypeStubs = false

[tool.coverage.run]
source = ["src/markit"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

